requirements.txt
discord.py==2.2.2
python-dotenv==1.0.0
asgiref==3.6.0
openai==0.27.4
revChatGPT==4.2.4
GoogleBard==0.0.7
EdgeGPT==0.1.25.1
supabase==1.0.3
weaviate-client==3.15.6
langchain==0.0.133